+++ packages/vanilla-components/src/components/dropdown/dropdown.vue
@@ -249,7 +249,7 @@ const doHide = async (): Promise<void> => {
       try {
         deactivate()
       }
-catch (e) {}
+      catch (e) {}
     }
     onHidden()
     focused.value = true