+++ packages/vanilla-components/src/core/config/common-classes.ts
@@ -43,7 +43,7 @@ const commonClasses = {
   buttonRadius: 'rounded-md',
   buttonSpacing: 'px-4 py-2',
   buttonShadow: 'shadow',
-  buttonText: 'sm:text-sm text-base font-medium leading-6 whitespace-nowrap',
+  buttonText: 'text-sm font-medium leading-6 whitespace-nowrap',
   buttonTransition: 'transition ease-in-out duration-15',
   buttonPosition: 'block justify-center inline-flex items-center',